+++ dwarfeng-datamark-core/src/test/java/com/dwarfeng/datamark/node/example/MultitonHandlerProcessExample.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.datamark.node.example;
 
 import com.dwarfeng.datamark.stack.handler.DatamarkHandler;
+import com.dwarfeng.dutil.basic.io.CT;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
@@ -31,23 +32,23 @@ public class MultitonHandlerProcessExample {
         Scanner scanner = new Scanner(System.in);
 
         // 显示欢迎信息。
-        System.out.println("开发者您好!");
-        System.out.println("这是一个示例, 用于演示 dwarfeng-datamark 的功能");
-        System.out.println("dwarfeng-datamark 是一个数据标记工具, 用于为数据提供当前时刻的标记");
-        System.out.println("为了更好的进行体验, 请您在运行本示例之前, 按照文档的说明创建配置文件");
+        CT.trace("开发者您好!");
+        CT.trace("这是一个示例, 用于演示 dwarfeng-datamark 的功能");
+        CT.trace("dwarfeng-datamark 是一个数据标记工具, 用于为数据提供当前时刻的标记");
+        CT.trace("为了更好的进行体验, 请您在运行本示例之前, 按照文档的说明创建配置文件");
         System.out.print("请按回车键继续...");
         scanner.nextLine();
 
         // 1. 获取当前的数据标记值。
-        System.out.println();
-        System.out.println("1. 获取当前的数据标记值...");
+        CT.trace("");
+        CT.trace("1. 获取当前的数据标记值...");
         try {
             String currentValue1 = instance1.get();
             String currentValue2 = instance2.get();
             String currentValue3 = instance3.get();
-            System.out.println("处理器 instance1 当前的数据标记值为: " + currentValue1);
-            System.out.println("处理器 instance2 当前的数据标记值为: " + currentValue2);
-            System.out.println("处理器 instance3 当前的数据标记值为: " + currentValue3);
+            CT.trace("处理器 instance1 当前的数据标记值为: " + currentValue1);
+            CT.trace("处理器 instance2 当前的数据标记值为: " + currentValue2);
+            CT.trace("处理器 instance3 当前的数据标记值为: " + currentValue3);
         } catch (Exception e) {
             LOGGER.warn("获取当前的数据标记值失败, 异常信息如下: ", e);
         }
@@ -55,19 +56,19 @@ public class MultitonHandlerProcessExample {
         scanner.nextLine();
 
         // 2. 刷新数据标记值。
-        System.out.println();
-        System.out.println("2. 刷新数据标记值...");
+        CT.trace("");
+        CT.trace("2. 刷新数据标记值...");
         try {
-            System.out.println("请编辑 ${datamark.resource.url} 对应的资源中的内容, 将其更改为新的数据标记值");
+            CT.trace("请编辑 ${datamark.resource.url} 对应的资源中的内容, 将其更改为新的数据标记值");
             System.out.print("请按回车键继续...");
             scanner.nextLine();
             String currentValue1 = instance1.refresh();
             String currentValue2 = instance2.refresh();
             String currentValue3 = instance3.refresh();
-            System.out.println("处理器 instance1 当前的数据标记值为: " + currentValue1);
-            System.out.println("处理器 instance2 当前的数据标记值为: " + currentValue2);
-            System.out.println("处理器 instance3 当前的数据标记值为: " + currentValue3);
-            System.out.println("请观察刷新后的数据标记值是否与您编辑后的数据标记值一致");
+            CT.trace("处理器 instance1 当前的数据标记值为: " + currentValue1);
+            CT.trace("处理器 instance2 当前的数据标记值为: " + currentValue2);
+            CT.trace("处理器 instance3 当前的数据标记值为: " + currentValue3);
+            CT.trace("请观察刷新后的数据标记值是否与您编辑后的数据标记值一致");
         } catch (Exception e) {
             LOGGER.warn("刷新并获取当前的数据标记值失败, 异常信息如下: ", e);
         }
@@ -75,22 +76,22 @@ public class MultitonHandlerProcessExample {
         scanner.nextLine();
 
         // 3. 更新数据标记值。
-        System.out.println();
-        System.out.println("3. 更新数据标记值...");
+        CT.trace("");
+        CT.trace("3. 更新数据标记值...");
         try {
-            System.out.println("请指定处理器 instance1 的新的数据标记值...");
+            CT.trace("请指定处理器 instance1 的新的数据标记值...");
             String neoValue1 = scanner.nextLine();
-            System.out.println("请指定处理器 instance2 的新的数据标记值...");
+            CT.trace("请指定处理器 instance2 的新的数据标记值...");
             String neoValue2 = scanner.nextLine();
-            System.out.println("请指定处理器 instance3 的新的数据标记值...");
+            CT.trace("请指定处理器 instance3 的新的数据标记值...");
             String neoValue3 = scanner.nextLine();
             String currentValue1 = instance1.update(neoValue1);
             String currentValue2 = instance2.update(neoValue2);
             String currentValue3 = instance3.update(neoValue3);
-            System.out.println("处理器 instance1 当前的数据标记值为: " + currentValue1);
-            System.out.println("处理器 instance2 当前的数据标记值为: " + currentValue2);
-            System.out.println("处理器 instance3 当前的数据标记值为: " + currentValue3);
-            System.out.println(
+            CT.trace("处理器 instance1 当前的数据标记值为: " + currentValue1);
+            CT.trace("处理器 instance2 当前的数据标记值为: " + currentValue2);
+            CT.trace("处理器 instance3 当前的数据标记值为: " + currentValue3);
+            CT.trace(
                     "请查看 ${datamark.resource.url} 对应的资源中的内容，观察是否与您指定的新的数据标记值一致"
             );
         } catch (Exception e) {
@@ -100,8 +101,8 @@ public class MultitonHandlerProcessExample {
         scanner.nextLine();
 
         // 显示结束信息。
-        System.out.println();
-        System.out.println("示例演示完毕, 感谢您测试与使用!");
+        CT.trace("");
+        CT.trace("示例演示完毕, 感谢您测试与使用!");
 
         ctx.stop();
         ctx.close();